// src/app/enums/days-of-month.enum.ts
export enum DaysOfMonth {
  ONE = 1,
  TWO = 2,
  THREE = 3,
  FOUR = 4,
  FIVE = 5,
  SIX = 6,
  SEVEN = 7,
  EIGHT = 8,
  NINE = 9,
  TEN = 10,
  ELEVEN = 11,
  TWELVE = 12,
  THIRTEEN = 13,
  FOURTEEN = 14,
  FIFTEEN = 15,
  SIXTEEN = 16,
  SEVENTEEN = 17,
  EIGHTEEN = 18,
  NINETEEN = 19,
  TWENTY = 20,
  TWENTY_ONE = 21,
  TWENTY_TWO = 22,
  TWENTY_THREE = 23,
  TWENTY_FOUR = 24,
  TWENTY_FIVE = 25,
  TWENTY_SIX = 26,
  TWENTY_SEVEN = 27,
  TWENTY_EIGHT = 28,
  TWENTY_NINE = 29,
  THIRTY = 30,
  THIRTY_ONE = 31
}

export const DAYS_OF_MONTH_OPTIONS = [
  { label: 'Ngày 1', value: DaysOfMonth.ONE },
  { label: 'Ngày 2', value: DaysOfMonth.TWO },
  { label: 'Ngày 3', value: DaysOfMonth.THREE },
  { label: 'Ngày 4', value: DaysOfMonth.FOUR },
  { label: 'Ngày 5', value: DaysOfMonth.FIVE },
  { label: 'Ngày 6', value: DaysOfMonth.SIX },
  { label: 'Ngày 7', value: DaysOfMonth.SEVEN },
  { label: 'Ngày 8', value: DaysOfMonth.EIGHT },
  { label: 'Ngày 9', value: DaysOfMonth.NINE },
  { label: 'Ngày 10', value: DaysOfMonth.TEN },
  { label: 'Ngày 11', value: DaysOfMonth.ELEVEN },
  { label: 'Ngày 12', value: DaysOfMonth.TWELVE },
  { label: 'Ngày 13', value: DaysOfMonth.THIRTEEN },
  { label: 'Ngày 14', value: DaysOfMonth.FOURTEEN },
  { label: 'Ngày 15', value: DaysOfMonth.FIFTEEN },
  { label: 'Ngày 16', value: DaysOfMonth.SIXTEEN },
  { label: 'Ngày 17', value: DaysOfMonth.SEVENTEEN },
  { label: 'Ngày 18', value: DaysOfMonth.EIGHTEEN },
  { label: 'Ngày 19', value: DaysOfMonth.NINETEEN },
  { label: 'Ngày 20', value: DaysOfMonth.TWENTY },
  { label: 'Ngày 21', value: DaysOfMonth.TWENTY_ONE },
  { label: 'Ngày 22', value: DaysOfMonth.TWENTY_TWO },
  { label: 'Ngày 23', value: DaysOfMonth.TWENTY_THREE },
  { label: 'Ngày 24', value: DaysOfMonth.TWENTY_FOUR },
  { label: 'Ngày 25', value: DaysOfMonth.TWENTY_FIVE },
  { label: 'Ngày 26', value: DaysOfMonth.TWENTY_SIX },
  { label: 'Ngày 27', value: DaysOfMonth.TWENTY_SEVEN },
  { label: 'Ngày 28', value: DaysOfMonth.TWENTY_EIGHT },
  { label: 'Ngày 29', value: DaysOfMonth.TWENTY_NINE },
  { label: 'Ngày 30', value: DaysOfMonth.THIRTY },
  { label: 'Ngày 31', value: DaysOfMonth.THIRTY_ONE }
];